Add GPI user interface, convert gpi_embed to user interface
A new environment variable was introduced `GPI_USERS` that specifies
libraries to load during GPI initialization. These libraries are GPI
users (or extensions of existing users). They use the spec
`lib_name:entry_func_name` to specify the library to load and entry
function to call. The entry function is a main()-like function.

gpi_embed was converted from the "embed" interface, which was a
single-user compile-time-linked user interface library, to one the new
GPI_USER interface. The "embed" interface was removed.

To accomplish what the previous embed interface provided, new GPI
functionality was added to allow the user to register callbacks to be
called when a simulator event occured, and when GPI shutdown was
occuring.