remove gpilog lib, move log registration to Python
gpilog made assumptions about the user using Python for their logging,
it was also an unnecessary additional module, only separated because it
had to link against Python.

Now, users can register any logger with the GPI, regardless of
implementation, and this does not require linking against the C API of
the implementation language.

Cocotb's Python-based logger was moved into Python, so that it can be
registered with the GPI on cocotb module initialization.