Add GPI user sim event register, remove embed interface, convert gpi_embed
This commit is intended to increase functionality for GPI extensions. It
accomplishes this by adding functions that allow a GPI user to register
callback functions for simulator startup, shutdown, and other event. While
this functionality is not of immediate use, if extensions ever wish to
receive arguments or react to shutdown, they will become necessary.

This new interface obliviates the need for the embed interface. The
embed interface only supports a single user and a single callback for
that user. The existing sole user of the embed interface, gpi_embed,
which is responsible for starting and killing the Python interpreter on
startup and shutdown, respectively, and loading its users, has been
converted to use this new interface.

The runtime simulator event register was added to the Python GPI binding
(simulatormodule), so that Python users can react to simulator events.
The startup and shutdown callbacks were not added because these are used
to start and kill the Python interpreter. This may change in the future.
The current simulator event handler in the embed interface is now moved
to the responsibility of the Python user.