Load custom entry point in Python from the GPI
gpi_embed no longer hardcodes to enter "cocotb:_initialize_testbench",
but uses an entry module and function specified by the
COCOTB_ENTRY_POINT environment variable. The variable is of the format:
"path.to.entry.module:entry_point.function". The GPI embed also assumes
the "_log_from_c", "_filter_from_c", and "_sim_event" function are
located in the entry module. This may be cleaned up in the future.

Additionally, the body of the "embed_sim_init" function has been
improved to remove some of the complexity regarding object lifetimes in
the function by using a "DEFER" statement, which is seen in a few other
languages.