Change how libraries are specified in GPI_EXTRA
Previously names were computed using "lib" + name + SHARED_OBJ_EXT.
This limited the libraries that could be loaded by GPI_EXTRA.

This was changed so no name computation takes place. This requires the
user to specify the full name of the library, but allows using paths in
the library name and loading libraries whose name doesn't start with
"lib".